Remove duplicate test and move test helper function out of actix-http-test
