+++ actix-http-test/CHANGES.md
@@ -1,7 +1,6 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-* Added `TestServer::get_negotiated_alpn_protocol` method.
 
 ## 3.0.0-beta.4 - 2021-04-02
 * Added `TestServer::client_headers` method. [#2097]
