+++ actix-http-test/Cargo.toml
@@ -27,7 +27,6 @@ default = []
 
 # openssl
 openssl = ["tls-openssl", "awc/openssl"]
-rustls = ["tls-rustls", "webpki"]
 
 [dependencies]
 actix-service = "2.0.0"
@@ -50,8 +49,6 @@ slab = "0.4"
 serde_urlencoded = "0.7"
 time = { version = "0.2.23", default-features = false, features = ["std"] }
 tls-openssl = { version = "0.10.9", package = "openssl", optional = true }
-tls-rustls = { version = "0.19", package = "rustls", optional = true }
-webpki = { version = "0.21.0", optional = true }
 
 [dev-dependencies]
 actix-web = { version = "4.0.0-beta.6", default-features = false, features = ["cookies"] }
