+++ actix-http/src/h2/service.rs
@@ -171,7 +171,7 @@ mod rustls {
             Error = TlsError<io::Error, DispatchError>,
             InitError = S::InitError,
         > {
-            let mut protos = vec!["h2".to_string().into()];
+            let mut protos = vec![b"h2".to_vec()];
             protos.extend_from_slice(&config.alpn_protocols);
             config.set_protocols(&protos);
